Int64

+++ src/cc/games/prices.cpp
@@ -44,14 +44,14 @@ int32_t prices_barupdate(struct prices_bar *bar,uint64_t pricebits)
     return(0);
 }
 
-int32_t prices_bardist(struct prices_bar *bar,uint32_t aveprice,uint64_t pricebits)
+int64_t prices_bardist(struct prices_bar *bar,uint32_t aveprice,uint64_t pricebits)
 {
-    int32_t dist = 0; uint32_t uprice = (uint32_t)pricebits;
+    int64_t dist = 0; uint32_t uprice = (uint32_t)pricebits;
     if ( aveprice != 0 )
     {
         dist = (uprice - aveprice);
         dist *= dist;
-        fprintf(stderr,"dist.%d (u %u - ave %u) %d\n",dist,uprice,aveprice,uprice-aveprice);
+        fprintf(stderr,"dist.%lld (u %u - ave %u) %d\n",(long long)dist,uprice,aveprice,uprice-aveprice);
     }
     return(dist);
 }
@@ -100,7 +100,7 @@ int64_t prices_blockinfo(int32_t height,char *acaddr)
                                 }
                             }
                             else if ( strcmp(acaddr,destaddr) != 0 )
-                                fprintf(stderr,"i.%d %.8f t%u %.4f v.%d %s lag.%d i.%d dist.%d\n",i,(double)tx.vout[0].nValue/COIN,timestamp,(double)uprice/10000,numvouts,destaddr,(int32_t)(pindex->nTime-timestamp),iter,prices_bardist(&refbar,aveprice,pricebits));
+                                fprintf(stderr,"i.%d %.8f t%u %.4f v.%d %s lag.%d i.%d dist.%lld\n",i,(double)tx.vout[0].nValue/COIN,timestamp,(double)uprice/10000,numvouts,destaddr,(int32_t)(pindex->nTime-timestamp),iter,(long long)prices_bardist(&refbar,aveprice,pricebits));
                         } else return(-3);
                     }
                 }
